+++ clients/vscode/src/commands/startProvider.ts
@@ -1,8 +1,0 @@
-import { IKosCommand } from './types';
-
-export const startProvider: IKosCommand = {
-  command: 'kos.startksp',
-  commandCallback: () => {
-    console.log('sup');
-  },
-};